refactor: consolidate kafka-common, fix circular dependency, and complete service setup
- Merged kafka-common and common-events into a single module
- Fixed SecurityAutoConfiguration cycle by removing ObjectProvider
- Changed monetary fields to BigDecimal for DB compatibility
- Added missing dependencies (Lombok, flyway-postgresql, spring-boot plugin)
- Configured application.yml and main classes for store-service & order-service
- Resolved Kafka hostname resolution (added to hosts)
- Added init-scripts for automatic database creation
- All services (auth, user, store, order) now build and start successfully

+++ order-service/src/main/java/com/orderprocessing/orderservice/model/Order.java
@@ -3,7 +3,7 @@ package com.orderprocessing.orderservice.model;
 import java.util.UUID;
 import java.time.Instant;
 import java.util.List;
-
+import java.math.BigDecimal;
 import jakarta.persistence.*;
 import lombok.Data;
 
@@ -28,7 +28,7 @@ public class Order {
     private Status status;
 
     @Column(name = "total_amount", nullable = false)
-    private double totalAmount;
+    private BigDecimal totalAmount;
 
     @Column(name = "created_at", nullable = false)
     private Instant createdAt;
