+++ store-service/src/main/java/com/orderprocessing/storeservice/dto/ProductDTO.java
@@ -2,7 +2,7 @@ package com.orderprocessing.storeservice.dto;
 
 import java.util.UUID;
 import java.time.Instant;
-
+import java.math.BigDecimal;
 import lombok.Data;
 
 @Data
@@ -10,7 +10,7 @@ public class ProductDTO {
     private UUID id;
     private String name;
     private String description;
-    private double price;
+    private BigDecimal price;
     private String category;
     private Instant createdAt;
     private Instant updatedAt;
